# apps/api/openapi.json
{
  "components" : {
    "schemas" : {
...
      "ApiResponseListTaskResponse" : {
        "properties" : {
          "data" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskResponse"
            },
            "type" : "array"
          },
          "meta" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "ApiResponseTaskDetailResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TaskDetailResponse"
          },
          "meta" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
      "ApiResponseTaskResponse" : {
        "properties" : {
          "data" : {
            "$ref" : "#/components/schemas/TaskResponse"
          },
          "meta" : {
            "additionalProperties" : {
              "type" : "object"
            },
            "type" : "object"
          }
        },
        "type" : "object"
      },
...
      "CreateSubtaskRequest" : {
        "properties" : {
          "title" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [
          "title"
        ],
        "type" : "object"
      },
      "CreateTaskRequest" : {
        "properties" : {
          "assigneeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "priority" : {
            "type" : "string"
          },
          "sectionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "title" : {
            "maxLength" : 500,
            "minLength" : 0,
            "type" : "string"
          }
        },
        "required" : [
          "title"
        ],
        "type" : "object"
      },
...
      "MoveTaskRequest" : {
        "properties" : {
          "afterTaskId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "beforeTaskId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sectionId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "Progress" : {
        "properties" : {
          "completed" : {
            "format" : "int32",
            "type" : "integer"
          },
          "total" : {
            "format" : "int32",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
...
      "PromoteTaskRequest" : {
        "properties" : {
          "sectionId" : {
            "format" : "uuid",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "TaskDetailResponse" : {
        "properties" : {
          "subtaskProgress" : {
            "$ref" : "#/components/schemas/Progress"
          },
          "subtasks" : {
            "items" : {
              "$ref" : "#/components/schemas/TaskResponse"
            },
            "type" : "array"
          },
          "task" : {
            "$ref" : "#/components/schemas/TaskResponse"
          }
        },
        "type" : "object"
      },
      "TaskResponse" : {
        "properties" : {
          "assigneeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "completed" : {
            "type" : "boolean"
          },
          "completedAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdAt" : {
            "format" : "date-time",
            "type" : "string"
          },
          "createdBy" : {
            "format" : "uuid",
            "type" : "string"
          },
          "description" : {
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "id" : {
            "format" : "uuid",
            "type" : "string"
          },
          "parentTaskId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "priority" : {
            "type" : "string"
          },
          "projectId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sectionId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "sortKey" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          },
          "updatedAt" : {
            "format" : "date-time",
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
      "UpdateTaskRequest" : {
        "properties" : {
          "assigneeId" : {
            "format" : "uuid",
            "type" : "string"
          },
          "completed" : {
            "type" : "boolean"
          },
          "description" : {
            "type" : "string"
          },
          "dueDate" : {
            "format" : "date",
            "type" : "string"
          },
          "priority" : {
            "type" : "string"
          },
          "title" : {
            "type" : "string"
          }
        },
        "type" : "object"
      },
...
    }
  },
  "info" : {
    "description" : "Cairn — self-hosted program & project management (M1).",
    "title" : "Cairn API",
    "version" : "v1"
  },
  "openapi" : "3.0.1",
  "paths" : {
...
    "/api/v1/projects/{projectId}/tasks" : {
      "get" : {
        "operationId" : "listByProject",
        "parameters" : [
          {
            "in" : "path",
            "name" : "projectId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "cursor",
            "required" : false,
            "schema" : {
              "type" : "string"
            }
          },
          {
            "in" : "query",
            "name" : "limit",
            "required" : false,
            "schema" : {
              "format" : "int32",
              "type" : "integer"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseListTaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      },
      "post" : {
        "operationId" : "create_3",
        "parameters" : [
          {
            "in" : "path",
            "name" : "projectId",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      }
    },
...
    "/api/v1/tasks/{id}" : {
      "delete" : {
        "operationId" : "delete_1",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      },
      "get" : {
        "operationId" : "get",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskDetailResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      },
      "patch" : {
        "operationId" : "update_2",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/UpdateTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      }
    },
    "/api/v1/tasks/{id}/move" : {
      "post" : {
        "operationId" : "move",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/MoveTaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      }
    },
    "/api/v1/tasks/{id}/promote" : {
      "post" : {
        "operationId" : "promote",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/PromoteTaskRequest"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      }
    },
    "/api/v1/tasks/{id}/subtasks" : {
      "post" : {
        "operationId" : "addSubtask",
        "parameters" : [
          {
            "in" : "path",
            "name" : "id",
            "required" : true,
            "schema" : {
              "format" : "uuid",
              "type" : "string"
            }
          }
        ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/CreateSubtaskRequest"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponseTaskResponse"
                }
              }
            },
            "description" : "OK"
          }
        },
        "tags" : [
          "task-controller"
        ]
      }
    },
...
  },
  "servers" : [
    {
      "url" : "/"
    }
  ]
}